move lamb_op to phi
