+++ paddle/fluid/eager/auto_code_generator/final_state_generator/python_c_gen.py
@@ -88,6 +88,8 @@ no_amp_list = [
     'rmsprop',
     'sgd_',
     'sgd',
+    'lamb_',
+    'lamb',
     'assign_value_',
     'sparse_momentum_',
     'sparse_momentum',
